[REF] html_editor, website: move ir_ui_view save methods to website
Following website refactoring [1], and web_editor deletion, models were
moved to html_editor [2].
This commit is moving `ir_ui_view` methods related to `save` to website,
as they are specific to the module.

[1]: https://github.com/odoo/odoo/commit/9fe45e2b7ddbbfd0445ffe25a859e67a316d02b2
[2]: https://github.com/odoo/odoo/commit/867d9a2f923a55c0bd6a281e2dea1f524fa04445

task-5065980
