+++ addons/html_editor/models/ir_ui_view.py
@@ -1,321 +1,16 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import copy
 import logging
-from lxml import etree, html
+from lxml import etree
 
-from odoo import api, models, _
-from odoo.exceptions import ValidationError, MissingError
-from odoo.addons.base.models.ir_ui_view import MOVABLE_BRANDING
+from odoo import api, models
+from odoo.exceptions import MissingError
 
 _logger = logging.getLogger(__name__)
-
-EDITING_ATTRIBUTES = MOVABLE_BRANDING + [
-    'data-oe-type',
-    'data-oe-expression',
-    'data-oe-translation-id',
-    'data-note-id'
-]
 
 
 class IrUiView(models.Model):
     _inherit = 'ir.ui.view'
-
-    def _get_cleaned_non_editing_attributes(self, attributes):
-        """
-        Returns a new mapping of attributes -> value without the parts that are
-        not meant to be saved (branding, editing classes, ...). Note that
-        classes are meant to be cleaned on the client side before saving as
-        mostly linked to the related options (so we are not supposed to know
-        which to remove here).
-
-        :param attributes: a mapping of attributes -> value
-        :return: a new mapping of attributes -> value
-        """
-        attributes = {k: v for k, v in attributes if k not in EDITING_ATTRIBUTES}
-        if 'class' in attributes:
-            classes = attributes['class'].split()
-            attributes['class'] = ' '.join([c for c in classes if c != 'o_editable'])
-        if attributes.get('contenteditable') == 'true':
-            del attributes['contenteditable']
-        return attributes
-
-    # ------------------------------------------------------
-    # Save from html
-    # ------------------------------------------------------
-
-    @api.model
-    def extract_embedded_fields(self, arch):
-        return arch.xpath('//*[@data-oe-model != "ir.ui.view"]')
-
-    @api.model
-    def extract_oe_structures(self, arch):
-        return arch.xpath('//*[hasclass("oe_structure")][contains(@id, "oe_structure")]')
-
-    @api.model
-    def get_default_lang_code(self):
-        return False
-
-    @api.model
-    def save_embedded_field(self, el):
-        Model = self.env[el.get('data-oe-model')]
-        field = el.get('data-oe-field')
-
-        model = 'ir.qweb.field.' + el.get('data-oe-type')
-        converter = self.env[model] if model in self.env else self.env['ir.qweb.field']
-
-        try:
-            value = converter.from_html(Model, Model._fields[field], el)
-            if value is not None:
-                # TODO: batch writes?
-                record = Model.browse(int(el.get('data-oe-id')))
-                if not self.env.context.get('lang') and self.get_default_lang_code():
-                    record.with_context(lang=self.get_default_lang_code()).write({field: value})
-                else:
-                    record.write({field: value})
-
-                if callable(Model._fields[field].translate):
-                    self._copy_custom_snippet_translations(record, field)
-
-        except (ValueError, TypeError):
-            raise ValidationError(_(
-                "Invalid field value for %(field_name)s: %(value)s",
-                field_name=Model._fields[field].string,
-                value=el.text_content().strip(),
-            ))
-
-    def save_oe_structure(self, el):
-        self.ensure_one()
-
-        if el.get('id') in self.key:
-            # Do not inherit if the oe_structure already has its own inheriting view
-            return False
-
-        arch = etree.Element('data')
-        xpath = etree.Element('xpath', expr="//*[hasclass('oe_structure')][@id='{}']".format(el.get('id')), position="replace")
-        arch.append(xpath)
-        attributes = self._get_cleaned_non_editing_attributes(el.attrib.items())
-        structure = etree.Element(el.tag, attrib=attributes)
-        structure.text = el.text
-        xpath.append(structure)
-        for child in el.iterchildren(tag=etree.Element):
-            structure.append(copy.deepcopy(child))
-
-        vals = {
-            'inherit_id': self.id,
-            'name': '%s (%s)' % (self.name, el.get('id')),
-            'arch': etree.tostring(arch, encoding='unicode'),
-            'key': '%s_%s' % (self.key, el.get('id')),
-            'type': 'qweb',
-            'mode': 'extension',
-        }
-        vals.update(self._save_oe_structure_hook())
-        oe_structure_view = self.env['ir.ui.view'].create(vals)
-        self._copy_custom_snippet_translations(oe_structure_view, 'arch_db')
-
-        return True
-
-    @api.model
-    def _copy_custom_snippet_translations(self, record, html_field):
-        """ Given a ``record`` and its HTML ``field``, detect any
-        usage of a custom snippet and copy its translations.
-        """
-        lang_value = record[html_field]
-        if not lang_value:
-            return
-
-        try:
-            tree = html.fromstring(lang_value)
-        except etree.ParserError as e:
-            raise ValidationError(str(e))
-
-        for custom_snippet_el in tree.xpath('//*[hasclass("s_custom_snippet")]'):
-            custom_snippet_name = custom_snippet_el.get('data-name')
-            custom_snippet_view = self.search([('name', '=', custom_snippet_name)], limit=1)
-            if custom_snippet_view:
-                self._copy_field_terms_translations(custom_snippet_view, 'arch_db', record, html_field)
-
-    @api.model
-    def _copy_field_terms_translations(self, records_from, name_field_from, record_to, name_field_to):
-        """ Copy model terms translations from ``records_from.name_field_from``
-        to ``record_to.name_field_to`` for all activated languages if the term
-        in ``record_to.name_field_to`` is untranslated (the term matches the
-        one in the current language).
-
-        For instance, copy the translations of a
-        ``product.template.html_description`` field to a ``ir.ui.view.arch_db``
-        field.
-
-        The method takes care of read and write access of both records/fields.
-        """
-        record_to.check_access('write')
-        field_from = records_from._fields[name_field_from]
-        field_to = record_to._fields[name_field_to]
-        record_to._check_field_access(field_to, 'write')
-
-        error_callable_msg = "'translate' property of field %r is not callable"
-        if not callable(field_from.translate):
-            raise TypeError(error_callable_msg % field_from)
-        if not callable(field_to.translate):
-            raise TypeError(error_callable_msg % field_to)
-        if not field_to.store:
-            raise ValueError("Field %r is not stored" % field_to)
-
-        # This will also implicitly check for `read` access rights
-        if not record_to[name_field_to] or not any(records_from.mapped(name_field_from)):
-            return
-
-        lang_env = self.env.lang or 'en_US'
-        langs = {lang for lang, _ in self.env['res.lang'].get_installed()}
-
-        # 1. Get translations
-        records_from.flush_model([name_field_from])
-        existing_translation_dictionary = field_to.get_translation_dictionary(
-            record_to[name_field_to],
-            {lang: record_to.with_context(prefetch_langs=True, lang=lang)[name_field_to] for lang in langs if lang != lang_env}
-        )
-        extra_translation_dictionary = {}
-        for record_from in records_from:
-            extra_translation_dictionary.update(field_from.get_translation_dictionary(
-                record_from[name_field_from],
-                {lang: record_from.with_context(prefetch_langs=True, lang=lang)[name_field_from] for lang in langs if lang != lang_env}
-            ))
-        for term, extra_translation_values in extra_translation_dictionary.items():
-            existing_translation_values = existing_translation_dictionary.setdefault(term, {})
-            # Update only default translation values that aren't customized by the user.
-            for lang, extra_translation in extra_translation_values.items():
-                if existing_translation_values.get(lang, term) == term:
-                    existing_translation_values[lang] = extra_translation
-        translation_dictionary = existing_translation_dictionary
-
-        # The `en_US` jsonb value should always be set, even if english is not
-        # installed. If we don't do this, the custom snippet `arch_db` will only
-        # have a `fr_BE` key but no `en_US` key.
-        langs.add('en_US')
-
-        # 2. Set translations
-        new_value = {
-            lang: field_to.translate(lambda term: translation_dictionary.get(term, {}).get(lang), record_to[name_field_to])
-            for lang in langs
-        }
-        record_to.env.cache.update_raw(record_to, field_to, [new_value], dirty=True)
-        # Call `write` to trigger compute etc (`modified()`)
-        record_to[name_field_to] = new_value[lang_env]
-
-    @api.model
-    def _save_oe_structure_hook(self):
-        return {}
-
-    @api.model
-    def _are_archs_equal(self, arch1, arch2):
-        # Note that comparing the strings would not be ok as attributes order
-        # must not be relevant
-        if arch1.tag != arch2.tag:
-            return False
-        if arch1.text != arch2.text:
-            return False
-        if arch1.tail != arch2.tail:
-            return False
-        if arch1.attrib != arch2.attrib:
-            return False
-        if len(arch1) != len(arch2):
-            return False
-        return all(self._are_archs_equal(arch1, arch2) for arch1, arch2 in zip(arch1, arch2))
-
-    @api.model
-    def _get_allowed_root_attrs(self):
-        return ['style', 'class', 'target', 'href']
-
-    def replace_arch_section(self, section_xpath, replacement, replace_tail=False):
-        # the root of the arch section shouldn't actually be replaced as it's
-        # not really editable itself, only the content truly is editable.
-        self.ensure_one()
-        arch = etree.fromstring(self.arch.encode('utf-8'))
-        # => get the replacement root
-        if not section_xpath:
-            root = arch
-        else:
-            # ensure there's only one match
-            [root] = arch.xpath(section_xpath)
-
-        root.text = replacement.text
-
-        # We need to replace some attrib for styles changes on the root element
-        for attribute in self._get_allowed_root_attrs():
-            if attribute in replacement.attrib:
-                root.attrib[attribute] = replacement.attrib[attribute]
-            elif attribute in root.attrib:
-                del root.attrib[attribute]
-
-        # Note: after a standard edition, the tail *must not* be replaced
-        if replace_tail:
-            root.tail = replacement.tail
-        # replace all children
-        del root[:]
-        for child in replacement:
-            root.append(copy.deepcopy(child))
-
-        return arch
-
-    @api.model
-    def to_field_ref(self, el):
-        # filter out meta-information inserted in the document
-        attributes = {k: v for k, v in el.attrib.items()
-                           if not k.startswith('data-oe-')}
-        attributes['t-field'] = el.get('data-oe-expression')
-
-        out = html.html_parser.makeelement(el.tag, attrib=attributes)
-        out.tail = el.tail
-        return out
-
-    @api.model
-    def to_empty_oe_structure(self, el):
-        out = html.html_parser.makeelement(el.tag, attrib=el.attrib)
-        out.tail = el.tail
-        return out
-
-    @api.model
-    def _set_noupdate(self):
-        self.sudo().mapped('model_data_id').write({'noupdate': True})
-
-    def save(self, value, xpath=None):
-        """ Update a view section. The view section may embed fields to write
-
-        Note that `self` record might not exist when saving an embed field
-
-        :param str xpath: valid xpath to the tag to replace
-        """
-        self.ensure_one()
-
-        arch_section = html.fromstring(
-            value, parser=html.HTMLParser(encoding='utf-8'))
-
-        if xpath is None:
-            # value is an embedded field on its own, not a view section
-            self.save_embedded_field(arch_section)
-            return
-
-        for el in self.extract_embedded_fields(arch_section):
-            self.save_embedded_field(el)
-
-            # transform embedded field back to t-field
-            el.getparent().replace(el, self.to_field_ref(el))
-
-        for el in self.extract_oe_structures(arch_section):
-            if self.save_oe_structure(el):
-                # empty oe_structure in parent view
-                empty = self.to_empty_oe_structure(el)
-                if el == arch_section:
-                    arch_section = empty
-                else:
-                    el.getparent().replace(el, empty)
-
-        new_arch = self.replace_arch_section(xpath, arch_section)
-        old_arch = etree.fromstring(self.arch.encode('utf-8'))
-        if not self._are_archs_equal(old_arch, new_arch):
-            self._set_noupdate()
-            self.write({'arch': etree.tostring(new_arch, encoding='unicode')})
-            self._copy_custom_snippet_translations(self, 'arch_db')
 
     @api.model
     def _view_get_inherited_children(self, view):
