+++ addons/html_editor/tests/test_views.py
@@ -30,34 +30,3 @@ class TestViews(TransactionCase):
         })
         # Test for RecursionError: maximum recursion depth exceeded in this function
         View._views_get(self.first_view)
-
-    def test_oe_structure_as_inherited_view(self):
-        View = self.env['ir.ui.view']
-
-        base = View.create({
-            'name': 'Test View oe_structure',
-            'type': 'qweb',
-            'arch': """<xpath expr='//t[@t-call="html_editor.test_first_view"]' position='after'>
-                        <div class="oe_structure" id='oe_structure_test_view_oe_structure'/>
-                    </xpath>""",
-            'key': 'html_editor.oe_structure_view',
-            'inherit_id': self.second_view.id
-        })
-
-        # check view mode
-        self.assertEqual(base.mode, 'extension')
-
-        # update content of the oe_structure
-        value = '''<div class="oe_structure" id="oe_structure_test_view_oe_structure" data-oe-id="%s"
-                         data-oe-xpath="/div" data-oe-model="ir.ui.view" data-oe-field="arch">
-                        <p>Hello World!</p>
-                   </div>''' % base.id
-
-        base.save(value=value, xpath='/xpath/div')
-
-        self.assertEqual(len(base.inherit_children_ids), 1)
-        self.assertEqual(base.inherit_children_ids.mode, 'extension')
-        self.assertIn(
-            '<p>Hello World!</p>',
-            base.inherit_children_ids.get_combined_arch(),
-        )
