+++ addons/website/static/src/builder/website_builder.js
@@ -42,6 +42,7 @@ import { MonetaryFieldPlugin } from "@html_builder/plugins/monetary_field_plugin
 import { Many2OneOptionPlugin } from "@html_builder/plugins/many2one_option_plugin";
 import { CustomizeTranslationTab } from "@website/builder/plugins/translation_tab/customize_translation_tab";
 import { CustomizeTranslationTabPlugin } from "./plugins/translation_tab/customize_translation_tab_plugin";
+import { WebsiteSavePlugin } from "@website/builder/plugins/website_save_plugin";
 
 const TRANSLATION_PLUGINS = [
     BuilderOptionsTranslationPlugin,
@@ -71,6 +72,7 @@ const TRANSLATION_PLUGINS = [
     MonetaryFieldPlugin,
     Many2OneOptionPlugin,
     CustomizeTranslationTabPlugin,
+    WebsiteSavePlugin,
 ];
 
 export class WebsiteBuilder extends Component {
